refactor: [#198] Chat 패키지 하위로ChatRoom, ChatMessage 관련 Test 이동
