+++ src/test/java/com/ssh/dartserver/domain/chat/UsersChatRoomApiTest.java
@@ -1,4 +1,4 @@
-package com.ssh.dartserver.domain.chatRoom;
+package com.ssh.dartserver.domain.chat;
 
 import com.ssh.dartserver.ApiTest;
 import com.ssh.dartserver.UserManager;
